+++ apps/example/flagship-code.config.ts
@@ -53,7 +53,6 @@ export default defineConfig({
    * @type {string[]}
    */
   plugins: [
-    '@brandingbrand/code-plugin-native-navigation',
     '@brandingbrand/code-plugin-asset',
     '@brandingbrand/code-plugin-app-icon',
     '@brandingbrand/code-plugin-splash-screen',