+++ apps/example/metro.config.js
@@ -4,9 +4,9 @@ const { getCacheVersion } = require('@brandingbrand/code-app-env/metro');
 
 /**
  * Metro configuration
- * https://facebook.github.io/metro/docs/configuration
+ * https://reactnative.dev/docs/metro
  *
- * @type {import('metro-config').MetroConfig}
+ * @type {import('@react-native/metro-config').MetroConfig}
  */
 const config = {
   cacheVersion: getCacheVersion(),