+++ apps/example/src/components/index.ts
@@ -1,0 +1,3 @@
+export {default as Header, type HeaderProps} from './Header';
+export {default as Section, type SectionProps} from './Section';
+export {Text, type TextProps} from './Text';