+++ packages/plugin-transform-template/src/transforms/0.79/index.ts
@@ -1,0 +1,8 @@
+import {transforms077} from '../0.77';
+
+import {default as iosAppDelegateTransform} from './ios/app-delegate-swift';
+
+export const transforms079 = {
+  ...transforms077,
+  iosAppDelegateTransform,
+};