+++ packages/plugin-verify-dependencies/src/profile/0.77.ts
@@ -24,6 +24,13 @@ export default {
     required: true,
   },
   /**
+   * React Native Navigation package configuration
+   * @property {string} version - Target version ^8.0.0
+   */
+  'react-native-navigation': {
+    version: '^8.0.0',
+  },
+  /**
    * React Native Babel preset configuration
    * @property {string} version - The semantic version requirement
    * @property {string[]} capabilities - Required Babel dependencies
@@ -43,4 +50,4 @@ export default {
     version: '^0.77.0',
     devOnly: true,
   },
-} satisfies Record<string, DependencyProfile>;
+} satisfies DependencyProfile;