+++ packages/plugin-verify-dependencies/CHANGELOG.md
@@ -1,5 +1,13 @@
 # @brandingbrand/code-plugin-verify-dependencies
 
+## 1.2.0
+
+### Minor Changes
+
+- 2fc57cc: Add ability to define a specific version to install when installed version is out of range
+- 83646f8: React Native 0.82 Support
+- c5f263b: React Native 0.81 Support
+
 ## 1.1.0
 
 ### Minor Changes